add skeleton of restful application
this is a great start that will submit a basic job
to the flux handle running alongside the server.
My next steps are to finish endpoints, work
on user interfaces, and then ensure we have
the automation we need. Ultimately I want to
deploy this alongside the Flux Operator as a
potential MiniCluster controller for jobs!

Signed-off-by: vsoch <vsoch@users.noreply.github.com>