add support for launcher
now that we have flux executors within workflow
managers, launching from a single job (e.g., in
the docker container) is likely not going to work.
We introduce here the idea of a launcher, or a known
command to run on the server, e.g., "nextflow" that
is able to launch flux jobs that are equivalently
owned by the instance. Currently the one issue
with this approach is the launcher writing to
output file instead of stdout/stderr, but I have
an issue open to figure out how to handle this.
This PR will also fix a current set of bugs with
parsing the checkbox (boolean) form fields - the
name attribute was in the wrong spot, doh.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>