start of work to add pam auth
we are currently treating a user account as basic auth, and validating with
pam. We will eventually want to instead encode with a shared secret. I figured
out the part of submitting a job by root on behalf of a user, next I need to
figure out how to reproduce flux python commands into command line stuff.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>