List jobs tweaks (#45)
* testing keeping flux-uri set
* get rid of mini
* print job spec
* refactor multi-user auth to not use pam (instead OAuth2 similar)
* submit needs to be string
* total refactor to use database

I am not happy with the current multi-user setup that tries to use pam,
and still is not using good practices like oauth2/jwt tokens. I also
do not like the idea of maintaining two modes of operation - one through
a flux user and one via sudo running the server and launching jobs as the
user. Instead, I think the flux restful server should be in charge of
authenticating users, and using a local database that can be created
by an entity like the flux operator, and then a secret to encrypt all
communication. I have a lot of work to update tests, examples, and docs,
and then I need to test with the flux operator, but this should at least
be the start of the crux of work. Next I will work on fixing up the
Python client to do the proper back and forth for an oauth2 token.

* working on testing
* update docs
* ensure we have secret key in testing container
* add migrations directory
* update for launcher case
* ensure we include completion fields

Signed-off-by: vsoch <vsoch@users.noreply.github.com>